+++ .idea/modules.xml
@@ -3,6 +3,7 @@
   <component name="ProjectModuleManager">
     <modules>
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/ninchatsdk/networkdispatcher/networkdispatcher.iml" filepath="$PROJECT_DIR$/.idea/modules/ninchatsdk/networkdispatcher/networkdispatcher.iml" group="ninchat-sdk-android/ninchatsdk/networkdispatcher" />
+      <module fileurl="file://$PROJECT_DIR$/.idea/modules/networkdispatcher2/networkdispatcher2.iml" filepath="$PROJECT_DIR$/.idea/modules/networkdispatcher2/networkdispatcher2.iml" group="ninchat-sdk-android/networkdispatcher2" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/ninchat-migration-ninchat-sdk-android.iml" filepath="$PROJECT_DIR$/.idea/modules/ninchat-migration-ninchat-sdk-android.iml" group="ninchat-sdk-android" />
       <module fileurl="file://$PROJECT_DIR$/.idea/modules/ninchatsdk/ninchat-migration-ninchat-sdk-android-ninchatsdk.iml" filepath="$PROJECT_DIR$/.idea/modules/ninchatsdk/ninchat-migration-ninchat-sdk-android-ninchatsdk.iml" group="ninchat-sdk-android/ninchatsdk" />
       <module fileurl="file://$PROJECT_DIR$/ninchat-ninchat-sdk-android.iml" filepath="$PROJECT_DIR$/ninchat-ninchat-sdk-android.iml" group="ninchat-sdk-android" />
