+++ networkdispatcher2/proguard-rules.pro
