+++ networkdispatcher2/src/main/java/com/ninchat/sdk/NinchatBeginICE.kt
@@ -1,0 +1,22 @@
+package com.ninchat.sdk
+
+import com.ninchat.client.Props
+import com.ninchat.client.Session
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.withContext
+
+class NinchatBeginICE {
+    companion object {
+        suspend fun execute(currentSession: Session? = null): Long =
+                withContext(Dispatchers.IO) {
+                    val params = Props()
+                    params.setString("action", "begin_ice")
+                    val actionId: Long = try {
+                        currentSession?.send(params, null) ?: -1
+                    } catch (e: Exception) {
+                        -1
+                    }
+                    actionId
+                }
+    }
+}
