--- ninchatsdk/networkdispatcher/src/main/AndroidManifest.xml
@@ -1,4 +1,0 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.networkdispatcher">
-    <uses-permission android:name="android.permission.INTERNET" />
-</manifest>
